feat(campaign): add campaign creation functionality and update model structure
